Fix re-render issue, fetch data on App load only
