+++ src/components/Missions/Missions.jsx
@@ -1,15 +1,9 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
+import React from 'react';
+import { useSelector } from 'react-redux';
 import MissionItem from './MissionItem';
 import NavBar from '../Layout/NavBar';
-import { getMissionData } from '../../redux/features/missions/missionSlice';
 
 const Missions = () => {
-  const dispatch = useDispatch();
-  useEffect(() => {
-    dispatch(getMissionData());
-  }, [dispatch]);
-
   const { missionItem, isLoading } = useSelector((store) => store.missions);
 
   if (isLoading) {
